requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
twilio==9.0.1
elevenlabs==1.2.3  # Use the latest version you're working with
starlette==0.37.2   # FastAPI dependency
websockets==12.0    # Required for WebSocket support
requests==2.31.0    # Used internally by many SDKs
